Cascade lors d'une suppression, demande de confirmation lors de la suppression d'un produit

+++ resources/js/Vue/Components/Pages/Products.vue
@@ -59,8 +59,11 @@ export default {
     },
     remove(id)
     {
-      productService.remove(id)
-        .then(() => this.fetchProducts());
+        if(!confirm('Supprimer ce produit ?'))
+            return ;
+
+        productService.remove(id)
+            .then(() => this.fetchProducts());
     }
   }
 }
